+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_query_operator_range.sql
@@ -24,6 +24,7 @@ Set local citus.log_remote_commands to on; -- Will print Citus rewrites of the q
 Set local citus.log_local_commands to on; -- Will print the local queries 
 Set local log_statement to 'all'; -- Verbose logging
 -- if there is no index we don't optimize the  a > 1 & a < 5 to  a <> {1, 5} (aka the range operator) 
+SET local documentdb_core.enableCollation TO on;
 SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filter": { "a": { "$gt": 1 }, "a" : {"$lt" : 5} }, "collation" : {"locale" : "en", "strength" : 1} }');
 EXPLAIN(costs off) SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filter": { "a": { "$gt": 1 }, "a" : {"$lt" : 5} }, "collation" : {"locale" : "en", "strength" : 1}  }');
 END;
@@ -49,6 +50,7 @@ SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filt
 EXPLAIN(costs off) SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filter": { "a": { "$gt": 1 }, "a" : {"$lt" : 5} } }');
 
 -- when collation is present $push down of $range query is not done, and we use the unoptimized version (workitem=3423305)
+SET local documentdb_core.enableCollation TO on;
 SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filter": { "a": { "$gt": 1 }, "a" : {"$lt" : 5} }, "collation" : {"locale" : "en", "strength" : 1} }');
 EXPLAIN(costs off) SELECT document FROM bson_aggregation_find('db', '{ "find": "range_query", "filter": { "a": { "$gt": 1 }, "a" : {"$lt" : 5} } ,"collation" : {"locale" : "en", "strength" : 1} }');
 END;