+++ pg_documentdb/include/aggregation/bson_query.h
@@ -23,7 +23,7 @@ typedef void (*ProcessQueryFilterFunc)(void *context);
 
 bool TraverseQueryDocumentAndGetId(bson_iter_t *queryDocument,
 								   bson_value_t *idValue, bool errorOnConflict,
-								   bool *hasNonIdFields);
+								   bool *hasNonIdFields, bool *isIdValueCollationAware);
 
 void TraverseQueryDocumentAndProcess(bson_iter_t *queryDocument, void *context,
 									 ProcessQueryValueFunc processValueFunc,